# swagger/crud.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "crud.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/items": {
      "get": {
        "operationId": "ListItems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemListItemsResponse"
            }
          }
        },
        "tags": [
          "ItemService"
        ]
      },
      "post": {
        "operationId": "CreateItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemCreateItemResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/itemItem"
            }
          }
        ],
        "tags": [
          "ItemService"
        ]
      },
      "put": {
        "operationId": "UpdateItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemUpdateItemResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/itemItem"
            }
          }
        ],
        "tags": [
          "ItemService"
        ]
      }
    },
    "/items/{id}": {
      "get": {
        "operationId": "GetItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemGetItemResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ItemService"
        ]
      },
      "delete": {
        "operationId": "DeleteItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemDeleteItemResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ItemService"
        ]
      }
    }
  },
  "definitions": {
    "itemCreateItemResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "itemDeleteItemResponse": {
      "type": "object"
    },
    "itemGetItemResponse": {
      "type": "object",
      "properties": {
        "item": {
          "$ref": "#/definitions/itemItem"
        }
      }
    },
    "itemItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "itemListItemsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/itemItem"
          }
        }
      }
    },
    "itemUpdateItemResponse": {
      "type": "object"
    }
  }
}